% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatCharting.R
\name{createBoxFunc}
\alias{createBoxFunc}
\title{Function to draw a box on the scene}
\usage{
createBoxFunc(topLeftX, topLeftY, width, height, fillColour = NA,
  borderColour = "black")
}
\arguments{
\item{topLeftX}{x co-ordinate of top left corner of box}

\item{topLeftY}{y co-ordinate of top left corner of box}

\item{width}{Width of the box}

\item{height}{Height of the box}

\item{fillColour}{Colour to fill the box in with}

\item{borderColour}{Colour of the box edge}
}
\description{
Function to draw a box on the scene
}

