% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatCharting.R
\name{createCircleFunc}
\alias{createCircleFunc}
\title{Function to draw a circle on the scene}
\usage{
createCircleFunc(centerX, centerY, radius, fillColour = NA,
  borderColour = "black")
}
\arguments{
\item{centerX}{x co-ordinate of the center of the circle}

\item{centerY}{y co-ordinate of the center of the circle}

\item{radius}{Radius of the circle}

\item{fillColour}{Colour to fill the circle in with}

\item{borderColour}{Colour of the circle edge}
}
\description{
Function to draw a circle on the scene
}

