% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatCharting.R
\name{createSceneFunc}
\alias{createSceneFunc}
\title{Function to create a blank canvas / scene for drawing objects onto later}
\usage{
createSceneFunc(bottomLeftX, bottomLeftY, width, height, main = "",
  xlab = "", ylab = "", ann = T, xaxt = NULL, yaxt = NULL,
  xlim = NULL, ylim = NULL, frame.plot = T)
}
\arguments{
\item{bottomLeftX}{Bottom left x co-ordinate of the canvas}

\item{bottomLeftY}{Bottom left y co-ordinate of the canvas}

\item{width}{Canvas width}

\item{height}{Canvas height}

\item{main}{Text title of the plot}

\item{xlab}{Label for x-axis}

\item{ylab}{Label for y-axis}

\item{ann}{See plot ann}

\item{xaxt}{See plot xaxt}

\item{yaxt}{See plot yaxt}

\item{xlim}{See plot xlim}

\item{ylim}{See plot ylim}

\item{frame.plot}{Bool to enable or disable drawing of a frame around the canvas}
}
\description{
Function to create a blank canvas / scene for drawing objects onto later
}

