% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neat.R
\name{NEATSimulation.GetStateHistoryForGenomeAndSpecies}
\alias{NEATSimulation.GetStateHistoryForGenomeAndSpecies}
\title{Runs a genome and tracks the state history (will run the most fit by default)}
\usage{
NEATSimulation.GetStateHistoryForGenomeAndSpecies(simulation, genomeNum = NA,
  speciesNum = NA)
}
\arguments{
\item{simulation}{Takes a NEATSimulation class}

\item{genomeNum}{the genome number to run}

\item{speciesNum}{the species number to run}
}
\value{
State history
}
\description{
Runs a genome and tracks the state history (will run the most fit by default)
}

