% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neat.R
\name{NEATSimulation.RunSingleGeneration}
\alias{NEATSimulation.RunSingleGeneration}
\title{Runs a single generation}
\usage{
NEATSimulation.RunSingleGeneration(simulation, createVideo = F,
  videoPath = "videos", videoName = "", framesPerSecond = 1)
}
\arguments{
\item{simulation}{Takes a NEATSimulation class}

\item{createVideo}{True/False to save a video of the highest fitness simulation}

\item{videoPath}{Path to where to save the video}

\item{videoName}{Name of the video}

\item{framesPerSecond}{The frames per second of the video}
}
\value{
NEATSimulation class with new generation of genomes
}
\description{
Takes in a simulation, runs all the genomes, evaluates fitness and breeds the new generation
}

