% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neat.R
\name{newConfigNEAT}
\alias{newConfigNEAT}
\title{Configuration for setting the number of system inputs/outputs, the max number of nodes and the total number of genomes}
\usage{
newConfigNEAT(numInputs, numOutputs, maxNumOfNodes, speciesPopulation = 200)
}
\arguments{
\item{numInputs}{The number of inputs to the neural network}

\item{numOutputs}{The number of outputs from the neural network}

\item{maxNumOfNodes}{The maximum number of neural network nodes}

\item{speciesPopulation}{The number of genomes to simulate}
}
\value{
configNEAT class
}
\description{
Configuration for setting the number of system inputs/outputs, the max number of nodes and the total number of genomes
}

